fixup! Use server broadcast events to send events via Mercure
