+++ appinfo/info.xml
@@ -21,6 +21,6 @@
 	<repository>https://github.com/nextcloud/push.git</repository>
 	<screenshot>https://raw.githubusercontent.com/nextcloud/push/master/screenshots/0.3.0.png</screenshot>
 	<dependencies>
-		<nextcloud min-version="18" max-version="18"/>
+		<nextcloud min-version="19" max-version="19"/>
 	</dependencies>
 </info>
